kuksa-client: Get rid of ActuatorDataEntry
The way we differentiate between sensor-only or actuator data entries is
currently flawed. The amount of extra logic required to get it right is
substantial as it would require querying entry type in several places
before get/set RPCs. For the little benefit the user gets in the end,
this seems unnecessary.