Changing check_push_rights.yml logic
We always use the default GITHUB_TOKEN for pushing.
Whether this has rughts to push to ghcr.io depends
on whether the code "originates" (i.e. push to branch
or internal PR) from our repo. As the login/build/push
workflows all reference the upstream eclipse project
we check for it.

Signed-off-by: Sebastian Schildt <sebastian.schildt@de.bosch.com>